# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Zinc Search engine API documents https://docs.zincsearch.com",
        "title": "Zinc Search engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Zinc Search",
            "url": "https://www.zincsearch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4080",
    "basePath": "/",
    "paths": {
        "/api/_analyze": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Analyze",
                "operationId": "Analyze",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/index.AnalyzeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/{index}/_analyze": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Analyze",
                "operationId": "AnalyzeIndex",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/index.AnalyzeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/{index}/_search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search V1",
                "operationId": "SearchV1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ZincQueryForSDK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/api/{index}/_settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get index settings",
                "operationId": "GetSettings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Set index Settings",
                "operationId": "SetSettings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.IndexSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "index.AnalyzeResponse": {
            "type": "object",
            "properties": {
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/index.AnalyzeResponseToken"
                    }
                }
            }
        },
        "index.AnalyzeResponseToken": {
            "type": "object",
            "properties": {
                "end_offset": {
                    "type": "integer"
                },
                "keyword": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "start_offset": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "meta.Analyzer": {
            "type": "object",
            "properties": {
                "char_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter": {
                    "description": "compatibility with es, alias for TokenFilter",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lowercase": {
                    "description": "for type=pattern",
                    "type": "boolean"
                },
                "pattern": {
                    "description": "for type=pattern",
                    "type": "string"
                },
                "stopwords": {
                    "description": "for type=pattern,standard,stop",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokenizer": {
                    "type": "string"
                },
                "type": {
                    "description": "options for compatible",
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meta.Highlight": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Highlight"
                    }
                },
                "fragment_size": {
                    "type": "integer"
                },
                "number_of_fragments": {
                    "type": "integer"
                },
                "post_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pre_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "meta.IndexAnalysis": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Analyzer"
                    }
                },
                "char_filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "filter": {
                    "description": "compatibility with es, alias for TokenFilter",
                    "type": "object",
                    "additionalProperties": true
                },
                "token_filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "tokenizer": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "meta.IndexSettings": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/meta.IndexAnalysis"
                },
                "number_of_replicas": {
                    "type": "integer"
                },
                "number_of_shards": {
                    "type": "integer"
                }
            }
        },
...
        "v1.AggregationBucket": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/v1.AggregationResponse"
                    }
                },
                "doc_count": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "v1.AggregationDateRange": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "v1.AggregationNumberRange": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "number"
                },
                "to": {
                    "type": "number"
                }
            }
        },
        "v1.AggregationParams": {
            "type": "object",
            "properties": {
                "agg_type": {
                    "type": "string"
                },
                "aggs": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/v1.AggregationParams"
                    }
                },
                "date_ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.AggregationDateRange"
                    }
                },
                "field": {
                    "type": "string"
                },
                "ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.AggregationNumberRange"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "weight_field": {
                    "description": "Field name to be used for setting weight for primary field for weighted average aggregation",
                    "type": "string"
                }
            }
        },
        "v1.AggregationResponse": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.AggregationBucket"
                    }
                },
                "value": {}
            }
        },
        "v1.Hit": {
            "type": "object",
            "properties": {
                "@timestamp": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                },
                "_index": {
                    "type": "string"
                },
                "_score": {
                    "type": "number"
                },
                "_source": {},
                "_type": {
                    "type": "string"
                },
                "highlight": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "v1.Hits": {
            "type": "object",
            "properties": {
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Hit"
                    }
                },
                "max_score": {
                    "type": "number"
                },
                "total": {
                    "$ref": "#/definitions/v1.Total"
                }
            }
        },
        "v1.QueryParams": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "term": {
                    "type": "string"
                },
                "terms": {
                    "description": "For multi phrase query",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "v1.SearchResponse": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/v1.AggregationResponse"
                    }
                },
                "error": {
                    "type": "string"
                },
                "hits": {
                    "$ref": "#/definitions/v1.Hits"
                },
                "timed_out": {
                    "type": "boolean"
                },
                "took": {
                    "description": "Time it took to generate the response",
                    "type": "integer"
                }
            }
        },
        "v1.Total": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Count of documents returned",
                    "type": "integer"
                }
            }
        },
        "v1.ZincQueryForSDK": {
            "type": "object",
            "properties": {
                "_source": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aggs": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/v1.AggregationParams"
                    }
                },
                "explain": {
                    "type": "boolean"
                },
                "from": {
                    "type": "integer"
                },
                "highlight": {
                    "$ref": "#/definitions/meta.Highlight"
                },
                "max_results": {
                    "type": "integer"
                },
                "query": {
                    "$ref": "#/definitions/v1.QueryParams"
                },
                "search_type": {
                    "description": "SearchType is the type of search to perform. Can be match, match_phrase, query_string, etc",
                    "type": "string"
                },
                "sort_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}